requirements.txt
openai>=1.0.0
httpx>=0.25.0
feedparser>=6.0.0
praw>=7.7.0
pydantic>=2.0.0
python-dotenv>=1.0.0
colorama>=0.4.6
flask>=3.0.0
flask-cors>=4.0.0
gunicorn>=21.0.0


